+++ inst/NEWS.md
@@ -1,3 +1,5 @@
+- New template based on van 't Veer and Giner-Sorollam (2016). See `?vantveer_prereg`.
+
 # prereg 0.2.0
 
 - New template based on AsPredicted.org questions. See `?aspredicted_prereg`.